Add files via upload
This is the Random Quote Generator App. It uses principles learnt in the You Are Awesome App to create a fun app that has an app icon, launchscreen color, launch screen image, a sound two-way binding option, and a section that shows the current date and time. And for the main part, it has a "generate quote" button that generates random quotes each time, also randomizing the sound each time! Amazing!